refactor: move root types to constants
