+++ src/utils/constants.ts
@@ -1,6 +1,8 @@
 import BigNumber from 'bignumber.js';
 import { TxTags } from 'polymesh-types/types';
 
+import { TransactionArgumentType } from '~/types';
+
 export const MAX_DECIMALS = 6;
 export const MAX_TICKER_LENGTH = 12;
 export const MAX_MODULE_LENGTH = 32;
@@ -24,3 +26,63 @@ export const MAX_CONCURRENT_REQUESTS = 200;
 export const TREASURY_MODULE_ADDRESS = 'modlpm/trsry';
 export const BATCH_REGEX = RegExp('(b|s?B)atch');
 export const DEFAULT_GQL_PAGE_SIZE = 25;
+
+const didTypes = ['IdentityId'];
+
+const addressTypes = [
+  'AccountId',
+  'AccountIdOf',
+  'LookupTarget',
+  'Address',
+  'AuthorityId',
+  'SessionKey',
+  'ValidatorId',
+  'AuthorityId',
+  'KeyType',
+  'SessionKey',
+];
+
+const balanceTypes = ['Amount', 'AssetOf', 'Balance', 'BalanceOf'];
+
+const numberTypes = ['u8', 'u16', 'u32', 'u64', 'u128', 'u256', 'U256', 'BlockNumber'];
+
+const textTypes = ['String', 'Text', 'Ticker'];
+
+const booleanTypes = ['bool'];
+
+const dateTypes = ['Moment'];
+
+const rootTypes: Record<
+  string,
+  | TransactionArgumentType.Did
+  | TransactionArgumentType.Address
+  | TransactionArgumentType.Balance
+  | TransactionArgumentType.Number
+  | TransactionArgumentType.Text
+  | TransactionArgumentType.Boolean
+  | TransactionArgumentType.Date
+> = {};
+
+didTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Did;
+});
+addressTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Address;
+});
+balanceTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Balance;
+});
+numberTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Number;
+});
+textTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Text;
+});
+booleanTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Boolean;
+});
+dateTypes.forEach(type => {
+  rootTypes[type] = TransactionArgumentType.Date;
+});
+
+export const ROOT_TYPES = rootTypes;
